IGNITE-24564 Remove redundant index column serializer.
