--- modules/catalog/src/main/java/org/apache/ignite/internal/catalog/descriptors/CatalogIndexColumnDescriptorSerializers.java
@@ -1,49 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.ignite.internal.catalog.descriptors;
-
-import java.io.IOException;
-import org.apache.ignite.internal.catalog.storage.serialization.CatalogObjectSerializer;
-import org.apache.ignite.internal.catalog.storage.serialization.CatalogSerializer;
-import org.apache.ignite.internal.util.io.IgniteDataInput;
-import org.apache.ignite.internal.util.io.IgniteDataOutput;
-
-/**
- * Serializers for {@link CatalogSortedIndexDescriptor}.
- */
-public class CatalogIndexColumnDescriptorSerializers {
-    /**
-     * Serializer for {@link CatalogSortedIndexDescriptor}.
-     */
-    @CatalogSerializer(version = 1, since = "3.0.0")
-    static class IndexColumnDescriptorSerializerV1 implements CatalogObjectSerializer<CatalogIndexColumnDescriptor> {
-        @Override
-        public CatalogIndexColumnDescriptor readFrom(IgniteDataInput input) throws IOException {
-            String name = input.readUTF();
-            CatalogColumnCollation collation = CatalogColumnCollation.unpack(input.readByte());
-
-            return new CatalogIndexColumnDescriptor(name, collation);
-        }
-
-        @Override
-        public void writeTo(CatalogIndexColumnDescriptor descriptor, IgniteDataOutput output) throws IOException {
-            output.writeUTF(descriptor.name());
-            output.writeByte(CatalogColumnCollation.pack(descriptor.collation()));
-        }
-    }
-}
