+++ modules/catalog/src/main/java/org/apache/ignite/internal/catalog/storage/DropSchemaSerializers.java
@@ -27,6 +27,7 @@ import org.apache.ignite.internal.util.io.IgniteDataOutput;
  * Serializers for {@link DropSchemaEntry}.
  */
 public class DropSchemaSerializers {
+
     /**
      * Serializer for {@link DropSchemaEntry}.
      */
